[docs] merge both custom render and overhaul example
- give this example its own folder, it's gotten complex enough at this point

- also give it its own custom toggles and tabs, since codesandbox doesn't generate a link with relative imports
